Deprecate one.alf.files and move functions to one.alf.path
