+++ one/registration.py
@@ -26,7 +26,7 @@ from iblutil.io import hashfile
 from iblutil.util import Bunch, ensure_list
 
 import one.alf.io as alfio
-from one.alf.files import session_path_parts, get_session_path, folder_parts, filename_parts
+from one.alf.path import session_path_parts, get_session_path, folder_parts, filename_parts
 from one.alf.spec import is_valid
 import one.alf.exceptions as alferr
 from one.api import ONE
